refactor: package 이동 및 추상 클래스 상속 변경
