+++ src/main/java/bowling/domain/frame/LastFrame.java
@@ -2,8 +2,8 @@ package bowling.domain.frame;
 
 import bowling.domain.Pins;
 import bowling.domain.score.Score;
-import bowling.domain.state.Miss;
-import bowling.domain.state.Ready;
+import bowling.domain.state.finish.Miss;
+import bowling.domain.state.running.Ready;
 import bowling.domain.state.State;
 
 import java.util.LinkedList;
