+++ src/main/java/bowling/domain/state/running/FirstBowl.java
@@ -1,10 +1,13 @@
-package bowling.domain.state;
+package bowling.domain.state.running;
 
 import bowling.domain.Pins;
 import bowling.domain.score.Score;
+import bowling.domain.state.finish.Miss;
+import bowling.domain.state.finish.Spare;
+import bowling.domain.state.State;
 import bowling.exception.InvalidPitchException;
 
-public class FirstBowl implements State {
+public class FirstBowl extends Running {
 
     private static final String GUTTER = "-";
